add/ remove user from project

+++ back/src/repositories/UserProjectRepository.ts
@@ -1,5 +1,6 @@
 import UsersProjects from "../models/UsersProjects";
 import User from "../models/User";
+import { error } from "console";
 
 class UserProjectRepository {
     async createUserProject(userId: number, projectId: number) {
@@ -28,24 +29,39 @@ class UserProjectRepository {
         });
     }
 
-    async deleteUserProject(projectId: number) {
-        return await UsersProjects.destroy({
-            where: {
-                ProjectId: projectId
-            }
-        });
-    }
-
-    async deleteUserProjects(userId: number) {
-        return await UsersProjects.destroy({
-            where: {
-                UserId: userId
-            }
-        });
-    }
-
     async findAll(filters: any = {}) {
         return await UsersProjects.findAll({ where: filters });
+    }
+
+    async addUsers(id: number, usersIds: number[]) {
+        const usersProjects = usersIds.map(userId => ({
+            ProjectId: id,
+            UserId: userId
+        }));
+        const isUsersIn = await UsersProjects.findAll({
+            where: {
+                ProjectId: id,
+                UserId: usersIds
+            }
+        })
+        if(isUsersIn.length > 0) throw new Error("User already in project")
+        return await UsersProjects.bulkCreate(usersProjects);
+    }
+    
+    async removeUsers(id: number, usersIds: number[]){
+        const isUsersIn = await UsersProjects.findAll({
+            where: {
+                ProjectId: id,
+                UserId: usersIds
+            }
+        })
+        if(isUsersIn.length == 0) throw new Error("User not in project")
+        return await UsersProjects.destroy({
+            where: {
+                ProjectId: id,
+                UserId: usersIds
+            }
+        });
     }
 }
